Use cvmrepo.s3.cern.ch as cvmfs package download url
This is strictly better and removes one point of failure.
Packages are downloaded anyway from s3.cern.ch, so it's better to
rely on this url than to introduce one more host.

Next step will be to update the cvmfs-release package with a failover to cvmrepo.web.cern.ch, then
this should be really robust.

A note on Amazonlinux 2023: this didn't work when I tested it with the amazonlinux docker image
as 2023 uses (micro-)dnf instead of yum. Even s/yum/dnf/ didn't work as micro-dnf doesn't seem to support
installing local packages (?). For the release package, also the redhat-release dependency was not supported.

I'll take a look into adding a yum repo for amazonlinux, if it can really just link to the centos7/9 packages that should
be very easy to do and let you remove this workaround.